Avoid calculation accuracy
Convert float to integer the truncation is required.

+++ Calculator.cs
@@ -5,11 +5,11 @@ namespace MyCalculator
 {
 	using System;
 	public class Calculator : ICalculator {
-		public double Add (double a, double b) {
-			return a + b;
+		public int Add (double a, double b) {
+			return (int)(a + b);
 		}
 
-		public double Sub (double a, double b) {
+		public int Sub (double a, double b) {
 			return Add (a, -b);
 		}
 	}
