+++ ICalculator.cs
@@ -2,7 +2,7 @@ namespace MyCalculator
 {
 	using System;
 	public interface ICalculator {
-		double Add (double a, double b);
-		double Sub (double a, double b);
+		int Add (double a, double b);
+		int Sub (double a, double b);
 	}
 }
